requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
langchain
torch
pypdf
python-docx
unstructured
